05_recommendation_engine/requirements.txt
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4
seaborn==0.13.2
scikit-learn==1.4.2
scikit-surprise==1.1.3
scipy==1.13.0
jupyter==1.0.0
